+++ src/modules/matrixGrid.module.ts
@@ -2,7 +2,7 @@ import {NgModule} from '@angular/core';
 
 import {ContentContainerSAComponent, ContentRowContainerSAComponent, FooterContainerSAComponent, FooterRowContainerSAComponent, GridContainerSAComponent, HeaderContainerSAComponent, HeaderRowContainerSAComponent, MatrixGridSAComponent} from '../components';
 import {ContentCellTemplateSADirective, ContentContainerTemplateSADirective, ContentRowContainerTemplateSADirective, FooterCellTemplateSADirective, FooterContainerTemplateSADirective, FooterRowContainerTemplateSADirective, GridContainerTemplateSADirective, HeaderCellTemplateSADirective, HeaderContainerTemplateSADirective, HeaderRowContainerTemplateSADirective, MatrixGridColumnSADirective, OrderableSADirective} from '../directives';
-import {CssGridTemplateColumnsSAPipe} from '../pipes';
+import {CssGridTemplateColumnsPipe} from '../pipes';
 
 /**
  * Module for matrix grid components and directives
@@ -31,7 +31,7 @@ import {CssGridTemplateColumnsSAPipe} from '../pipes';
         HeaderRowContainerTemplateSADirective,
         MatrixGridColumnSADirective,
         OrderableSADirective,
-        CssGridTemplateColumnsSAPipe,
+        CssGridTemplateColumnsPipe,
     ],
     exports:
     [
@@ -55,7 +55,7 @@ import {CssGridTemplateColumnsSAPipe} from '../pipes';
         HeaderRowContainerTemplateSADirective,
         MatrixGridColumnSADirective,
         OrderableSADirective,
-        CssGridTemplateColumnsSAPipe,
+        CssGridTemplateColumnsPipe,
     ],
 })
 export class MatrixGridModule