+++ src/pipes/cssGridTemplateColumns/cssGridTemplateColumns.pipe.ts
@@ -5,14 +5,14 @@ import {MatrixGridColumn} from '../../interfaces';
 /**
  * Creates css value for 'grid-template-columns' property
  */
-@Pipe({name: 'cssGridTemplateColumns', standalone: true,})
-export class CssGridTemplateColumnsSAPipe implements PipeTransform
+@Pipe({name: 'cssGridTemplateColumns'})
+export class CssGridTemplateColumnsPipe implements PipeTransform
 {
     /**
      * Creates css value for 'grid-template-columns' property
      * @param value - Array of columns that will be rendered
      */
-    public transform(value: MatrixGridColumn[]): string 
+    public transform(value: MatrixGridColumn[]): string
     {
         return value.map(col => col.width).join(' ');
     }