+++ src/plugins/contentRenderer/bodyHeader/table/header/basic/tableHeaderContentRenderer.component.ts
@@ -1,11 +1,11 @@
 import {Component, ChangeDetectionStrategy, Inject, Optional, HostBinding} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {MergeCssClassesSAPipe} from '@anglr/common';
+import {MergeCssClassesPipe} from '@anglr/common';
 
 import {TableHeaderContentRendererOptions} from '../../tableContentRenderer.interface';
 import {HeaderContentRendererAbstractComponent} from '../../../headerContentRendererAbstract.component';
 import {HEADER_CONTENT_RENDERER_OPTIONS} from '../../../../../../misc/tokens';
-import {CellContextSAPipe} from '../../../../../../pipes';
+import {CellContextPipe} from '../../../../../../pipes';
 import {provideCellContextFactoryFn} from '../../../../../../misc/providers';
 import {cellContextFactory} from '../../../../../../misc/utils';
 import {OrderableSADirective} from '../../../../../../directives';
@@ -35,8 +35,8 @@ const defaultOptions: TableHeaderContentRendererOptions =
     imports:
     [
         CommonModule,
-        CellContextSAPipe,
-        MergeCssClassesSAPipe,
+        CellContextPipe,
+        MergeCssClassesPipe,
         OrderableSADirective,
     ],
     providers: